--- babel.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  presets: [
-    ['@babel/preset-env', {
-      useBuiltIns: 'usage',
-      corejs: '3',
-    }],
-    '@babel/preset-react',
-  ],
-};
