+++ package.json
@@ -5,28 +5,18 @@
   "main": "src/index.js",
   "scripts": {
     "prebuild": "rm -fr ./dist",
-    "build": "webpack --mode production",
-    "start": "webpack serve --mode development",
+    "build": "rspack build --mode production",
+    "start": "rspack serve",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.17.0",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
     "@relative-ci/agent": "^4.1.3",
-    "babel-loader": "^8.2.3",
-    "core-js": "^3.21.0",
-    "css-loader": "^6.6.0",
+    "@rspack/cli": "^0.1.0",
     "eslint": "^8.8.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.5.3",
-    "webpack": "^5.68.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+    "eslint-config-airbnb": "^19.0.4"
   },
   "dependencies": {
     "antd": "^4.18.6",
