+++ src/index.html
@@ -5,6 +5,6 @@
     <title><%= htmlWebpackPlugin.options.title %></title>
   </head>
   <body>
-    <div id="root" />
+    <div id="root"></div>
   </body>
 </html>
